requirements.txt
aiohttp==3.8.1
asyncio==3.4.3
faiss-cpu==1.7.2
huggingface-hub==0.13.0
joblib==1.1.0
langchain==0.0.35
openai==0.27.0
pandas==1.4.1
psycopg2-binary==2.9.3
requests==2.27.1
sentence-transformers==2.1.0
SQLAlchemy==1.4.31
streamlit==1.8.1